+++ core/src/main/java/net/ivpn/core/v2/map/MapView.kt
@@ -96,7 +96,7 @@ class MapView @JvmOverloads constructor(
         }
 
         override fun onScroll(
-            e1: MotionEvent,
+            e1: MotionEvent?,
             e2: MotionEvent,
             distanceX: Float,
             distanceY: Float
@@ -109,8 +109,10 @@ class MapView @JvmOverloads constructor(
         }
 
         override fun onFling(
-                e1: MotionEvent, e2: MotionEvent,
-                velocityX: Float, velocityY: Float
+            e1: MotionEvent?,
+            e2: MotionEvent,
+            velocityX: Float,
+            velocityY: Float
         ): Boolean {
             fling((-velocityX).toInt(), (-velocityY).toInt())
             return true