+++ core/src/main/java/net/ivpn/core/v2/viewmodel/ConnectionViewModel.kt
@@ -35,8 +35,8 @@ import net.ivpn.core.common.session.SessionController
 import net.ivpn.core.common.utils.ConnectivityUtil
 import net.ivpn.core.common.utils.StringUtil
 import net.ivpn.core.rest.Responses
+import net.ivpn.core.rest.data.session.SessionErrorResponse
 import net.ivpn.core.rest.data.session.SessionNewResponse
-import net.ivpn.core.rest.data.wireguard.ErrorResponse
 import net.ivpn.core.v2.connect.createSession.ConnectionNavigator
 import net.ivpn.core.v2.connect.createSession.ConnectionState
 import net.ivpn.core.v2.dialog.Dialogs
@@ -257,7 +257,7 @@ class ConnectionViewModel @Inject constructor(
         }
     }
 
-    override fun onCreateError(throwable: Throwable?, errorResponse: ErrorResponse?) {
+    override fun onCreateError(throwable: Throwable?, errorResponse: SessionErrorResponse?) {
         if (!ConnectivityUtil.isOnline(context)) {
             navigator?.openErrorDialog(Dialogs.CONNECTION_ERROR)
             return
@@ -276,7 +276,7 @@ class ConnectionViewModel @Inject constructor(
                 navigator?.logout()
             }
             Responses.SESSION_TOO_MANY -> {
-                navigator?.openSessionLimitReachedDialogue()
+                navigator?.openSessionLimitReachedDialogue(errorResponse)
             }
         }
     }
@@ -284,11 +284,14 @@ class ConnectionViewModel @Inject constructor(
     override fun onUpdateSuccess() {
     }
 
-    override fun onUpdateError(throwable: Throwable?, errorResponse: ErrorResponse?) {
+    override fun onUpdateError(throwable: Throwable?, errorResponse: SessionErrorResponse?) {
         errorResponse?.let {
             if (it.status == Responses.SESSION_NOT_FOUND){
                 navigator?.logout()
             }
         }
+    }
+
+    override fun onDeviceLoggedOut() {
     }
 }