+++ store/src/main/java/net/ivpn/client/billing/BillingManagerWrapper.java
@@ -45,8 +45,8 @@ import net.ivpn.core.rest.data.addfunds.AddFundsRequestBody;
 import net.ivpn.core.rest.data.addfunds.AddFundsResponse;
 import net.ivpn.core.rest.data.addfunds.InitialPaymentRequestBody;
 import net.ivpn.core.rest.data.addfunds.InitialPaymentResponse;
+import net.ivpn.core.rest.data.session.SessionErrorResponse;
 import net.ivpn.core.rest.data.session.SessionNewResponse;
-import net.ivpn.core.rest.data.wireguard.ErrorResponse;
 import net.ivpn.core.rest.requests.common.Request;
 import net.ivpn.core.rest.requests.common.RequestWrapper;
 
@@ -60,12 +60,15 @@ import java.util.List;
 
 import javax.inject.Inject;
 
+import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.ADD_FUNDS_ERROR;
 import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.CREATE_SESSION;
 import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.CREATE_SESSION_ERROR;
 import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.INITIAL_PAYMENT;
 import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.INITIAL_PAYMENT_ERROR;
 import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.UPDATE_SESSION;
 import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.UPDATE_SESSION_ERROR;
+import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.PAYMENT_PENDING;
+import static net.ivpn.client.billing.BillingManagerWrapper.PurchaseState.INITIAL_PAYMENT_PENDING;
 
 @BillingScope
 public class BillingManagerWrapper {
@@ -130,8 +133,20 @@ public class BillingManagerWrapper {
                             && ConsumableProducts.INSTANCE.getConsumableSKUs().contains(purchase.getProducts().get(0))) {
                         billingManager.consumePurchase(purchase);
                     }
+
+                    if (purchase.getPurchaseState() == Purchase.PurchaseState.PENDING) {
+                        String sessionToken = userPreference.getSessionToken();
+                        if (sessionToken.isEmpty()) {
+                            setPurchaseState(INITIAL_PAYMENT_PENDING);
+                        } else {
+                            setPurchaseState(PAYMENT_PENDING);
+                        }
+                    }
+
+                    if (purchase.getPurchaseState() == Purchase.PurchaseState.PURCHASED) {
+                        startValidatingActivity(purchase);
+                    }
                 }
-                startValidatingActivity(purchases.get(0));
             }
 
             @Override
@@ -248,7 +263,7 @@ public class BillingManagerWrapper {
             }
 
             @Override
-            public void onCreateError(@Nullable Throwable throwable, @Nullable ErrorResponse errorResponse) {
+            public void onCreateError(@Nullable Throwable throwable, @Nullable SessionErrorResponse errorResponse) {
                 LOGGER.info("On create session Error: " + throwable + "/n" + errorResponse);
                 sessionController.unSubscribe(this);
                 setPurchaseState(CREATE_SESSION_ERROR);
@@ -268,18 +283,18 @@ public class BillingManagerWrapper {
                 if (response.getStatus() == Responses.SUCCESS) {
                     updateSession();
                 } else {
-                    setPurchaseState(INITIAL_PAYMENT_ERROR);
+                    setPurchaseState(ADD_FUNDS_ERROR);
                 }
             }
 
             @Override
             public void onError(Throwable throwable) {
-                setPurchaseState(INITIAL_PAYMENT_ERROR);
+                setPurchaseState(ADD_FUNDS_ERROR);
             }
 
             @Override
             public void onError(String string) {
-                setPurchaseState(INITIAL_PAYMENT_ERROR);
+                setPurchaseState(ADD_FUNDS_ERROR);
             }
         });
     }
@@ -297,7 +312,7 @@ public class BillingManagerWrapper {
             }
 
             @Override
-            public void onUpdateError(@Nullable Throwable throwable, @Nullable ErrorResponse errorResponse) {
+            public void onUpdateError(@Nullable Throwable throwable, @Nullable SessionErrorResponse errorResponse) {
                 LOGGER.info("On create session Error: " + throwable + "/n" + errorResponse);
                 sessionController.unSubscribe(this);
                 setPurchaseState(UPDATE_SESSION_ERROR);
@@ -346,6 +361,8 @@ public class BillingManagerWrapper {
         CREATE_SESSION_ERROR,
         UPDATE_SESSION,
         UPDATE_SESSION_ERROR,
-        ADD_FUNDS_ERROR
+        ADD_FUNDS_ERROR,
+        PAYMENT_PENDING,
+        INITIAL_PAYMENT_PENDING
     }
 }